# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Main Z11 api",
        "description": "This is our main Z11 api. It contains all method to handle Z11 project...",
        "termsOfService": "",
        "contact": {
            "email": "kienlv58@gmail.com"
        },
        "license": {
            "name": "Private License",
            "url": "URL to the license"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/chapter/get_all/{take}/{skip}": {
            "get": {
                "tags": [
                    "5.Chapter"
                ],
                "summary": "get all Chapter",
                "description": "return chapter with take and skip",
                "operationId": "Chapter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "take",
                        "in": "path",
                        "description": "take from ....",
                        "required": true,
                        "type": "integer",
                        "default": "all"
                    },
                    {
                        "name": "skip",
                        "in": "path",
                        "description": "skip from",
                        "required": true,
                        "type": "integer",
                        "default": "0"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid value"
                    }
                }
            }
        },
...
        "/package/get_all/{take}/{skip}": {
            "get": {
                "tags": [
                    "4.Package"
                ],
                "summary": "get all package",
                "description": "return package with take and skip",
                "operationId": "package",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "take",
                        "in": "path",
                        "description": "take from ....",
                        "required": true,
                        "type": "integer",
                        "default": "all"
                    },
                    {
                        "name": "skip",
                        "in": "path",
                        "description": "skip from",
                        "required": true,
                        "type": "integer",
                        "default": "0"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid value"
                    }
                }
            }
        },
...
        "/purchase/get_user/{id}": {
            "get": {
                "tags": [
                    "Purchase"
                ],
                "summary": "get_user",
                "description": "get_user purchas with user_id",
                "operationId": "get_user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "formData",
                        "description": "user_id",
                        "required": true,
                        "schema": {
                            "required": [
                                "grant_type"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get succes"
                    },
                    "400": {
                        "description": "Invalid param"
                    }
                }
            }
        },
        "/purchase/getUserPurchase_package/{id}": {
            "get": {
                "tags": [
                    "Purchase"
                ],
                "summary": "getUserPurchase_package",
                "description": "getUserPurchase_package purchas with user_id",
                "operationId": "get_user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "formData",
                        "description": "user_id",
                        "required": true,
                        "schema": {
                            "required": [
                                "grant_type"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get succes"
                    },
                    "400": {
                        "description": "Invalid param"
                    }
                }
            }
        },
        "/purchase/getUserPurchase_explain/{id}": {
            "get": {
                "tags": [
                    "Purchase"
                ],
                "summary": "getUserPurchase_explain",
                "description": "getUserPurchase_explain purchas with user_id",
                "operationId": "get_user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "formData",
                        "description": "user_id",
                        "required": true,
                        "schema": {
                            "required": [
                                "grant_type"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get succes"
                    },
                    "400": {
                        "description": "Invalid param"
                    }
                }
            }
        },
        "/purchase/getPurchaseId/{id}": {
            "get": {
                "tags": [
                    "Purchase"
                ],
                "summary": "getPurchaseId",
                "description": "getPurchaseId purchas with purchase_id",
                "operationId": "getPurchaseId",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "purchase_id",
                        "in": "formData",
                        "description": "purchase_id",
                        "required": true,
                        "schema": {
                            "required": [
                                "purchase_id"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get succes"
                    },
                    "400": {
                        "description": "Invalid param"
                    }
                }
            }
        },
        "/purchase/get_all/{take}/{skip}": {
            "get": {
                "tags": [
                    "Purchase"
                ],
                "summary": "get all Chapter",
                "description": "return purchase with take and skip",
                "operationId": "purchase",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "take",
                        "in": "path",
                        "description": "take from ....",
                        "required": true,
                        "type": "integer",
                        "default": "all"
                    },
                    {
                        "name": "skip",
                        "in": "path",
                        "description": "skip from",
                        "required": true,
                        "type": "integer",
                        "default": "0"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid value"
                    }
                }
            }
        },
        "/purchase/add_payment": {
            "post": {
                "tags": [
                    "Purchase"
                ],
                "summary": "add new purchase",
                "description": "add new purchase",
                "operationId": "purchaseadd",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "formData",
                        "description": "user_id ",
                        "required": true,
                        "schema": {
                            "required": [
                                "user_id"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    },
                    {
                        "name": "item_id",
                        "in": "formData",
                        "description": "item_id of value",
                        "required": true,
                        "schema": {
                            "required": [
                                "item_id"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    },
                    {
                        "name": "item_code",
                        "in": "formData",
                        "description": "select package or explain",
                        "required": true,
                        "schema": {
                            "required": [
                                "text_value"
                            ],
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "add succes"
                    },
                    "400": {
                        "description": "Invalid Value"
                    }
                }
            }
        },
        "/purchase/delete": {
            "post": {
                "tags": [
                    "Purchase"
                ],
                "summary": "delete purchase ",
                "description": "delete with purchase_id",
                "operationId": "purchasedelete",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "uid",
                        "in": "formData",
                        "description": "uid delete",
                        "required": true,
                        "schema": {
                            "required": [
                                "uid"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    },
                    {
                        "name": "purchase_id",
                        "in": "formData",
                        "description": "purchase_id",
                        "required": true,
                        "schema": {
                            "required": [
                                "purchase_id"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delete succes"
                    },
                    "400": {
                        "description": "Invalid Value"
                    }
                }
            }
        },
...
    },
...
}